+++ subsys/usb/common/CMakeLists.txt
@@ -1,4 +1,7 @@
 # SPDX-FileCopyrightText: Copyright Nordic Semiconductor ASA
 # SPDX-License-Identifier: Apache-2.0
 
-zephyr_include_directories(include)
+if(CONFIG_USBD_VIDEO_CLASS)
+  zephyr_include_directories(.)
+  zephyr_sources(uvc.c)
+endif()
